Update patch set 4
Patch Set 4:

There are really two cases I'm thinking about.  1. We use boringSSL in some test code that is loaded with dlopen.  2. BoringSSL is in the Android source tree, and the plugin we work on uses it.

1.  The test harness mimics a low library that gives access to stuff in the Android Trust Zone (TEE) and is loaded using dlopen/dlclose because that is what the real library uses.  We can't easily stop using dlopen, but we can try to make it a long term goal.  If you don't put in the patch, it will probably be easier for us to just fork a local copy of boringSSL just for testing.

2. The mediadrm does not use dlopen/dlclose directly to load boringSSL.  It is an android plugin and statically links boringSSL.  I don't know how many other android plugins also use boringSSL.  In Android O, on the device I tested, I was able to reproduce the problem, so I assume the android plugin system used something like dlopen/dlclose.  I have figured out that on this device, the plugin was not "binderized".  That means it is not using the new android HIDL interface to load and unload the plugin.  I will try to run my test app on a device with a binderized plugin to see if the bug goes away.

In either case, it would be easier for me if you included the fix and we just mirrored you.  In case 1, I can work toward removing dlopen/dlclose but it would not be in production until Android Q.  In case 2, I don't have control of the android source, but I can let them know about the bug and let them decided the severity of the problem.

Patch-set: 4
Reviewer: Gerrit User 8853 <8853@66b711a4-aa9b-3dac-9062-916540830212>
Label: Commit-Queue=0