Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> There are really two cases I'm thinking about.  1. We use boringSSL in some test code that is loaded with dlopen.  2. BoringSSL is in the Android source tree, and the plugin we work on uses it.
> 
> 1.  The test harness mimics a low library that gives access to stuff in the Android Trust Zone (TEE) and is loaded using dlopen/dlclose because that is what the real library uses.  We can't easily stop using dlopen, but we can try to make it a long term goal.  If you don't put in the patch, it will probably be easier for us to just fork a local copy of boringSSL just for testing.

dlopen isn't a problem, dlclose is. Lots of libraries will not handle dlclose gracefully and, as noted, absl simply banned it. The usual pattern is to lazily dlopen as needed and then leave the library in memory. (In time, the code and rodata pages will be dropped by the kernel anyway.)

> In either case, it would be easier for me if you included the fix and we just mirrored you.

You fix doesn't fix the memory leaks, just some of the memory leaks, and only those leaks that would be solved by running code in a thread and exiting the thread when done. It sounds like the leak of the thread-local key is what's causing immediate issues and, if it's useful, we'll land a fix for that.

But I don't think the rest of the patch carries its weight. If you absolutely must fix some subset of the memory leaks from dlclose, then I guess you'll need to fork. You shouldn't, but nor can I stop you.

Patch-set: 4
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>