Freeze RSA public key too.
While freeze the private key, we canonicalise the width of |rsa->n|.
However, there can be concurrent public-key operations occuring on the
key at the same time.

Therefore this change adds the act of freezing the public-key. This
means that |rsa->n| is known to be minimal-width and will no longer
change.

Change-Id: I6d3ab9bd3673059f83d23edae219b97b6d50defb