Send the fake session ID in the test suite.
NSS only enables compatibility mode on the server if the client
requested it by way of the session ID. This is slightly off as a client
has no way not to request it when offering a TLS 1.2 session, but it is
in the spec.

So our tests are usable for other stacks, send a fake session ID in the
runner by default. The existing EmptySessionID asserts that BoringSSL
behaves as we expect it to. The intent is that NSS will disable that
test but can otherwise leave the rest enabled.

Change-Id: I370bf90aba1805c2f6970ceee0d29ecf199f437d