Require getrandom in FIPS mode.
Previously we supported polling /dev/urandom for sufficient entropy to
meet the FIPS requirements. This change removes support for that and
always requires getrandom support in FIPS mode. The getrandom syscall
implements a requirement that 128 bits of entropy be collected before
any is supplied to user-space, eliminating the need to handle this
ourselves.

Change-Id: Ie4d1217fd89e15fd3d73598fb5a47466ec24b6e7
Update-Note: we already have this behaviour internally. We should update the local patch with one that uses our internal message because that contains a more useful link.