Write runner and shim transcripts to two files, and combine later.
This is prefactoring for a coming change to the shim that will write
handoff and handback messages (which are serialized SSLConnection
objects) to the transcript.

This causes interleaving of writes from the shim and writes from the
runner, which wasn't a problem before.  To address this, the runner's
writes are moved to a temporary file, which is appended to the tagged
data written by the shim after the test completes.

Change-Id: Iae34d28ec1addfe3ec4f3c77008248fe5530687c