Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > WSAStartup happens globally in crypto/test/gtest_main.h these days, because 
> > all these test files are linked into a single GTest binary.
> 
> Not if one uses Google's official GoogleTest package and
> not the one under 'third_party/googletest'. 
> I must be a complete retard to use the official GoogleTest package 
> (which has no call to WSAStartup()).

Using an external GTest should hopefully work. It is, indeed, how most downstream consumers set things up. Using the external gtest_main is not going to work and does not generally work. GTest provides a default gtest_main target, but it is not uncommon for tests to need different setup logic, which is why the gtest_main target is merely a convenience hook, with InitGoogleTest + RUN_ALL_TESTS being the actual API.

https://github.com/google/googletest/blob/master/googletest/docs/Primer.md#writing-the-main-function

Patch-set: 3