Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (3 comments)
> 
> What's the motivation here? You should be using ALPN, not NPN, at this point.

Overall motivation is that despite NPN deprecation many servers continue to expose lists of supported protocols over NPN. Because ALPN requires the server instead of the client to pick a protocol, it doesn't offer a way for a server to enumerate a list of supported protocols other than as an oracle function. Consequently if one is curious about what servers support what protocols, NPN is the only way to do so. I acknowledge there will be a limited time window of utility for doing so, but that window is open at present.

For instance, it was only via this mechanism that I was able to see that www.google.com supports grpc-exp as a valid protocol.

Patch-set: 1
Reviewer: Gerrit User 8969 <8969@66b711a4-aa9b-3dac-9062-916540830212>