Fix bssl client/server's error-handling.
Rather than printing the SSL_ERROR_* constants, print the actual error.
This should be a bit more understandable. Debugging this also uncovered
an issue with WSAGetLastError on Windows and thread locals.  Finally,
listen for FD_CLOSE along with FD_READ on Windows. Connection close does
not signal FD_READ.

(The select loop only barely works on Windows anyway due to issues with
stdin and line buffering, but if we take stdin out of the equation,
FD_CLOSE can be tested.)

Change-Id: If991259915acc96606a314fbe795fe6ea1e295e8