Update patch set 2
Patch Set 2:

IMO it is not great to have this test depend on `RAND_bytes()` as `RAND_bytes()` is actually a much more complicated function. `RAND_bytes()` is also nondeterministic. In theory `RAND_bytes()` could even depend on `CRYPTO_memcmp()` in its implementation (I don't know whether it currently does or not).

Patch-set: 2
CC: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>