Remove bsaes-x86_64.pl.
In the case where BSAES is used, the "no_hw" table-based implementation
of AES block encryption will be used. In particular, the last block of
an input of a length that isn't divisible by 16 bytes will be encrypted
using the "no_hw" implementation. The "no_hw" implementation isn't
side-channel free because it uses lookup tables.

The CPU requirements are exactly the same for BSAES and VPAES (SSE3).
If AES-NI is available then AES-NI will be used in preference to both
BSAES and VPAES. Thus, both are fallbacks for platforms that are
becoming increasingly rare. The performance advantage for
multiple-block operations that BSAES has over VPAES is thus
increasingly less important.

More generally, this reduces the amount of code that needs to be
tested, fuzzed, understood, and otherwise maintained by a
non-trivial amount. It also makes 32-bit x86 and x86-64 more similar
since x86 only has VPAES and doesn't have BSAES.

Change-Id: Ia691c45dcc33e50a960d6abb16fbe84c1366657f