Remove bsaes-x86_64.pl.
When using the same AES_KEY for multiple AES modes (regardless of
whether it is a good idea) or through different APIs, it isn't clear
that correct results will always be obtained. (Long ago, I experienced
confusing results when using the AES code in uninteneded ways, though
I've forgotten the details.)

In the case of x86-64 BSAES vs VPAES in particular, note that the CPU
requirements are exactly the same for both in e_aes.c. It seems BSAES
is used for some modes and VPAES is used for other modes. If AES-NI is
available then it is intended that AES-NI be used in preference to both
BSAES and VPAES. Thus, both are fallbacks for platforms that are
becoming increasingly rare. The performance advantage for
multiple-block operations that BSAES has over VPAES is thus
increasingly less important. Given that, it makes sense to just remove
the BSAES code on x86-64. This will reduce the amount of code that
needs to be examined to ensure it is always OK to use any given AES_KEY
for any mode of operation through any internal or external API.

Also, this avoids us needing to examine whether the fallback logic for
BSAES mode to handle smaller inputs will trigger the best--fastest,
most likely to be side-channel-free--single-block implementation, at
least for x86-64.

More generally, this reduces the amount of code that needs to be
tested, fuzzed, and otherwise maintained by a non-trivial amount.

Change-Id: Ia691c45dcc33e50a960d6abb16fbe84c1366657f